requirements.txt
flask
pycryptodome
djongo
jsonfield
django-sslserver
